Stop conflating same-bitwidth integer types
Until now, VeriFast conflated integer types with the same known
bitwidth. For example, when specifying `-target LP64`, types `long`,
`long long`, `int64_t`, and `intptr_t` were conflated, and predicate
`llong_integer` was used to denote objects of any of these types.

As of now, these types are all considered distinct, and distinct
predicates are used for them. For this purpose, predicates
`long_integer`, `ulong_integer`, `intptr`, and `uintptr` have been
introduced.

This breaking change is a step towards making VeriFast sound with
respect to the fact that, in the absence of the -fno-strict-aliasing
flag, some C compilers (including GCC and Clang, but not MSVC) perform
type-based alias analysis. Such analysis assumes that pointers to
distinct types do not alias. GCC and Clang consider some pairs of types
to be distinct for this purpose even though they have the same bitwidth.
In particular, it appears that they always consider the standard integer
types `char`, `short`, `int`, `long`, and `long long` to be distinct,
even though, e.g. on Linux/AMD64, `long` and `long long` are both 8
bytes wide. See https://godbolt.org/z/5zccMc4Ka . It also appears that the
extended integer types, such as `int64_t` and `intptr_t`, are mapped to
standard integer types, so that on Linux/AMD64, `intptr_t` and `int64_t`
are conflated with `long`, and therefore considered distinct from `long
long`. See https://godbolt.org/z/8v7345abb.

However, considering the arbitrary and undocumented nature of how
extended integer types are mapped to standard integer types, VeriFast
currently considers all of these types to be distinct. This has the
benefit that as of now, the same predicate is used for a given type,
regardless of the -target setting.

Still, it seems potentially useful to support the option of making
VeriFast conflate particular sets of same-bitwidth types. If such a
feature is desired, please let me know.

+++ examples/crypto_ccs/protocols/enc_and_hmac/enc_and_hmac.c
@@ -32,7 +32,7 @@ void sender(char *enc_key, char *hmac_key, char *msg, unsigned int msg_len)
   int socket;
   havege_state havege_state;
 
-  unsigned int iv_off = 0;
+  size_t iv_off = 0;
   char iv[16];
   char hmac[64];
   aes_context aes_context;
@@ -140,7 +140,7 @@ int receiver(char *enc_key, char *hmac_key, char *msg)
   int enc_size;
   char hmac[64];
   unsigned int temp;
-  unsigned int iv_off = 0;
+  size_t iv_off = 0;
   char iv[16];
   aes_context aes_context;
